+++ .github/workflows/publish.yml
@@ -39,8 +39,8 @@ jobs:
       - name: Build and push
         uses: docker/build-push-action@v5
         with:
-          context: ./gh_archive_flow
-          file: ./gh_archive_flow/Dockerfile
+          context: ./airflow
+          file: ./airflow/Dockerfile
           push: true
           tags: |
             ${{ env.REGISTRY }}/${{ env.IMAGE_ORG }}/${{ env.IMAGE_NAME }}:${{ steps.vars.outputs.branch }}-${{ steps.vars.outputs.sha_short }}
