+++ airflow/.dockerignore
